Move updatable group types to /models
