+++ src/models/group/updatable-group.ts
@@ -6,13 +6,10 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-export * from './creatable-group';
-export * from './group';
-export * from './is-group';
-export * from './raw-creatable-group';
-export * from './raw-group';
-export * from './raw-updatable-group';
-export * from './to-group';
-export * from './to-raw-creatable-group';
-export * from './to-raw-updatable-group';
-export * from './updatable-group';
+import { NumericID } from '../../value-objects';
+
+export interface UpdatableGroup {
+	id: NumericID;
+	name?: string;
+	description?: string | null;
+}
